CLN: renamed utilities.py to data_source_utilities.py
